+++ lib/doorkeeper/models/concerns/revocable.rb
@@ -8,6 +8,27 @@ module Doorkeeper
       def revoked?
         !!(revoked_at && revoked_at <= Time.now)
       end
+
+      def revoke_in(time)
+        update_attribute :revoked_at, Time.now + time
+      end
+
+      def revoke_previous_refresh_token!
+        if old_refresh_token && !old_refresh_token.revoked?
+          old_refresh_token.revoke_in(refresh_token_revoked_in)
+        end
+        if previous_refresh_token.present?
+          update_attribute :previous_refresh_token, ""
+        end
+      end
+
+      def old_refresh_token
+        @old_refresh_token ||= AccessToken.by_refresh_token(previous_refresh_token)
+      end
+
+      def refresh_token_revoked_in
+        Doorkeeper.configuration.refresh_token_revoked_in
+      end
     end
   end
 end