+++ spec/lib/models/revocable_spec.rb
@@ -33,4 +33,26 @@ describe 'Revocable' do
       expect(subject).not_to be_revoked
     end
   end
+
+  describe :revoke_previous_refresh_token! do
+    subject { FactoryGirl.build(:access_token, previous_refresh_token: 'old_refresh_token') }
+    previous_token = FactoryGirl.build(:access_token, token: "previous_token")
+    after { subject.revoke_previous_refresh_token! }
+
+    it 'revokes the previous token if present and sets the attribute :previous_refresh_token to ""' do
+      expect(Doorkeeper::AccessToken).to receive(:by_refresh_token).with(subject.previous_refresh_token).and_return(previous_token)
+      expect(previous_token).to receive(:revoke_in)
+      expect(subject).to receive(:update_attribute).with(:previous_refresh_token, "")
+    end
+
+    it 'does nothing if the previous refresh token is eq to ""' do
+      subject.previous_refresh_token = ""
+      expect(subject).to_not receive(:update_attribute).with(:previous_refresh_token, "")
+    end
+
+    it 'sets the attribute :previous_refresh_token to nil if the previous refresh token does not exist' do
+      expect(Doorkeeper::AccessToken).to receive(:by_refresh_token).with(subject.previous_refresh_token).and_return(nil)
+      expect(subject).to receive(:update_attribute).with(:previous_refresh_token, "")
+    end
+  end
 end